move gql to graphql file
