+++ contacts-app/src/api/index.js
@@ -1,1 +1,1 @@
-export * from './graphql'
+export * from './rest-api'
