+++ contacts-app/src/hooks/use-contacts.js
@@ -1,17 +1,7 @@
 import { useEffect, useState } from "react"
 import { useContactsContext } from "."
 import { gql, useQuery } from "@apollo/client"
-
-export const GET_CONTACTS = gql`
-  query {
-    contact {
-      nodes {
-        id
-        name
-      }
-    }
-  }
-`
+import { GET_CONTACTS } from "../api/graphql"
 
 function useContacts() {
   const { loading, error, data} = useQuery(GET_CONTACTS)
@@ -21,6 +11,22 @@ function useContacts() {
     data : {
       items :data ? data.contact.nodes : []
     }
+  }
+}
+
+function useContactsRest() {
+  const { state: data, loadContacts } = useContactsContext()
+  const [loading, setLoading] = useState(true)
+
+  useEffect(() => {
+    loadContacts().then(() => {
+      setLoading(false)
+    })
+  }, [])
+
+  return {
+    loading,
+    data
   }
 }
