+++ contacts-app/src/hooks/use-update-contact.js
@@ -1,16 +1,5 @@
 import { gql, useMutation } from "@apollo/client"
-import { GET_CONTACTS } from "./use-contacts"
-import { GET_CONTACT } from "./use-contact"
-
-const UPDATE_CONTACT = gql`
-  mutation updateContact($id: Int!, $name: String!, $mobileNumber: String){
-    updateContact(input: { id: $id, name: $name, mobileNumber: $mobileNumber }) {
-      id
-      name
-      mobileNumber
-    }
-  }
-`
+import { GET_CONTACT, GET_CONTACTS, UPDATE_CONTACT } from "../api/graphql"
 
 function useUpdateContact() {
   const [update, { data, loading, error }] = useMutation(UPDATE_CONTACT, {
